+++ Obsidian/Net/Packets/Play/Serverbound/UseItemPacket.cs
@@ -1,4 +1,5 @@
-﻿using Obsidian.Entities;
+﻿using Obsidian.API.Events;
+using Obsidian.Entities;
 using Obsidian.Serialization.Attributes;
 
 namespace Obsidian.Net.Packets.Play.Serverbound;
@@ -13,5 +14,22 @@ public partial class UseItemPacket : IServerboundPacket
 
     public int Id => 0x31;
 
-    public ValueTask HandleAsync(Server server, Player player) => ValueTask.CompletedTask;
+    public async ValueTask HandleAsync(Server server, Player player)
+    {
+        switch (Hand)
+        {
+            case Hand.MainHand:
+                await server.Events.InvokePlayerInteractAsync(new PlayerInteractEventArgs(player)
+                {
+                    Item = player.GetHeldItem()
+                });
+                break;
+            case Hand.OffHand:
+                await server.Events.InvokePlayerInteractAsync(new PlayerInteractEventArgs(player)
+                {
+                    Item = player.GetOffHandItem()
+                });
+                break;
+        }
+    }
 }
